Revert folder structure
