--- src/index.ts
@@ -1,2 +1,0 @@
-export { defineField } from './defineField';
-export type { InferFieldShape, InferFieldSchema } from './types';
